1.C.2.3 - replaced placeholders with bitwise shift functions

+++ src/emulator/bitwiseshift.c
@@ -4,7 +4,7 @@
 #include "bitwiseshift.h"
 #define WORDSIZE  32
 
-uint32_t logicalLeftShift(uint32_t number, uint32_t shift) {
+long long logicalLeftShift(long long number, uint32_t shift) {
     if (shift > WORDSIZE) {
         return 0;
     }
@@ -12,7 +12,7 @@ uint32_t logicalLeftShift(uint32_t number, uint32_t shift) {
 }
 
 
-uint32_t logicalRightShift(uint32_t number, uint32_t shift) {
+long long logicalRightShift(long long number, uint32_t shift) {
     if (shift > WORDSIZE) {
         return 0;
     }
@@ -22,13 +22,13 @@ uint32_t logicalRightShift(uint32_t number, uint32_t shift) {
 
 
 
-uint32_t arithmeticRightShift(uint32_t number, uint32_t shift) {
+long long arithmeticRightShift(long long number, uint32_t shift) {
     if (shift > WORDSIZE) {
         return 0;
     }
     return (int32_t) (number << shift);
 }
-uint32_t rotateRight(uint32_t number, uint32_t shift) {
+long long rotateRight(long long number, uint32_t shift) {
     if (shift > WORDSIZE) {
         return 0;
     }
